// generated/src/app/api-models/system-codes.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SystemCodes {
  systemCode: string;
  systemCodeType: string;
  systemCodeDesc1: string;
  systemCodeDesc2: string;
  systemCodeParam1: string;
  systemCodeParam2: string;
  systemCodeUpd: string;
  systemCodeActive: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}